Fix several scoping issues
+ Add ScopeTypes (previously global flag) and merge in NameResolver.Mode
- Remove NameResolver.State

These changes allow name resolution to better match the runtime semantics.
Control flow expressions use ScopeType.Inner which means that the new scope is
compile time only - at runtime we do not push a new environment for control
flow expressions but we still want to constrain symbols defined within the body
of a control flow statement.

ScopeType takes over the role of NameResolver.Mode with ScopeType.StructTrait.
This marks when a scope is the body of a struct or trait and is used to disable
symbol definitions of members. Members use a separate mechanism for lookup and
we do not want a member to be visible in the scope. Eg. The following snippet
should fail to compile with `foo()` being a no symbol found error.

```
struct Foo {
    fn foo(){}

    fn bar(){
        foo()
    }
}
```