Fix closure capture at top level
I'm not particularly convinced separating RefGlobal from RefUpvalue is
the right direction but for now that's what we're doing. This patch adds
a check to convert RefId (generated primarily by Set) to RefGlobal.

I would like to use RefUpvalue instead, however the resolution logic in
the interpreter assumes all RefUpvalues are locals of some environment
and will fail when trying to access globals.